fix(toxav): harden video processing and fix large frame handling
- Remove invalid vpx_img_free call in vc_iterate.
- Make vc_reconfigure_encoder atomic and add resolution limits.
- Support 32-bit RTPMessage lengths to correctly handle large frames.
- Ensure len is correctly set for assembled video frames.
- Add security checks for incoming frame sizes and resolutions.
- Significantly expand video and rtp unit tests covering security scenarios.
- Clean up VCSession structure and improve internal API consistency.
- Use #define for security constants in video module.

+++ toxav/rtp.h
@@ -57,7 +57,7 @@ typedef struct RTPSession RTPSession;
 
 /* RTPMessage accessors */
 const uint8_t *rtp_message_data(const RTPMessage *msg);
-uint16_t rtp_message_len(const RTPMessage *msg);
+uint32_t rtp_message_len(const RTPMessage *msg);
 uint8_t rtp_message_pt(const RTPMessage *msg);
 uint16_t rtp_message_sequnum(const RTPMessage *msg);
 uint64_t rtp_message_flags(const RTPMessage *msg);
